relocate overview screen
